requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-property==0.2.2
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-celery-results==2.5.1
django-filter==23.3
djangorestframework==3.14.0
flake8==6.1.0
kombu==5.3.4
Markdown==3.5.1
mccabe==0.7.0
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.10
